RL-183 Fix tests
