+++ src/test/java/gov/ca/cwds/rest/IntegratedResourceTestSuiteIT.java
@@ -1,7 +1,6 @@
 package gov.ca.cwds.rest;
 
 import gov.ca.cwds.rest.resources.ScreeningParticipantResourceIRT;
-import gov.ca.cwds.rest.resources.ScreeningRelationshipsWithCandidatesIRT;
 import gov.ca.cwds.rest.resources.ScreeningResourceIRT;
 import org.junit.ClassRule;
 import org.junit.runner.RunWith;
@@ -28,7 +27,7 @@ import gov.ca.cwds.test.support.BaseDropwizardApplication;
     ScreeningIntakeResourceIRT.class, ParticipantIntakeApiResourceIRT.class,
     ScreeningRelationshipResourceIRT.class, ContactResourceIRT.class,
     ScreeningParticipantResourceIRT.class,
-    ScreeningResourceIRT.class, ScreeningRelationshipsWithCandidatesIRT.class})
+    ScreeningResourceIRT.class})
 public class IntegratedResourceTestSuiteIT {
 
   @ClassRule
